requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
channels>=4.0.0
django-cors-headers>=4.1.0
drf-spectacular>=0.26.0
drf-nested-routers>=0.93.4
Pillow>=10.0.0
django-filter>=23.2
requests>=2.31.0
pytz>=2023.3
djangorestframework-simplejwt>=5.3.0
whitenoise>=6.5.0
uvicorn>=0.23.2
gunicorn>=21.2.0
channels-redis>=4.1.0
pytest>=7.3.1
pytest-django>=4.5.2